Add missing healthchecks
The two healthchecks added are required before we can make Whitehall
continuously deployed. As per the compatibility document [1]:

✅ Check for connectivity to Redis.
✅ Check for connectivity to MySQL.
✅ Missing check for connectivity to Memcached.
✅ Missing check for connectivity to AWS S3.

The S3 check simply makes a connection and searches for a S3 bucket
which doesn't exist. If the query is successful, then `nil` will be
returned, otherwise an error will be raised.

[1]: https://github.com/alphagov/govuk-rfcs/blob/main/rfc-128/compatibility.md

---

Trello:
https://trello.com/c/LCR0Am73/2541-5-enable-continuous-deployment-for-whitehall